Rename & fix Palermo Airport
The Palermo Airport is accessible by train from the Punta Raisi stations, recently reopened since 2015.
This merge requests renames “Punta Raisi” to identify it with the Airport of Palermo.
This station is also made parent of “Palermo Airport” (used only for buses), which has been also renamed to avoid confusions between train station and bus stop. The bus stop is no more suggestable, but will be included when looking for trips to his parent station.