Update Bern and Geneva IDs
According to SBB data, the Geneva city ID is 8501008, while Geneva Cornavin is 8587057.
In our database, Geneve Cornavin used the code for Geneve, Geneve Voyageur used the code of Geneve Cornavin and nobody used the code for Geneve Voyageurs.

This MR adds the right codes to Geneve, Geneve Cornavin and Geneve Voyageurs.