Rename San Marino to San Marino (Valbrenta) and remove translations
This PR renames San Marino to San Marino (Valbrenta) (in order to avoid confusion, as you may know a little but ancient Republic with the same name in the middle of Italy).
Also, removes the translations "Province of Vincenza" as sometimes consumers prefer to use the translation as the station name instead of the proper station name if present.