Fix CFF / PAO discrepancy for Genève
The main Genève train station is 5335 for CFF and Trenitalia, while it was 5336 for PAO.
In reality, station 5336 points to the bus stop outside of the main train station. By using station 5336 as departure or origin, CFF will return you a walk segment to get into the train station.
This PR puts all vendors ID against 5335 in order to avoid considering a transit in Genève main station as a change of stations.