Fix London St Pancras International name
Stations.csv contains two entries for London St Pancras International, marked as "same", but with names that differ slightly since a recent PR. This causes the station to potentially be duplicated when the file is consumed.

This PR unifies the names of the two entries as `London St Pancras International` (no dashes, no dots) to match how it is displayed by most retailers and rail related services.