Fix Ourense/Orense
The main station which used to have Empalme as a subtext isn't officially
named like that. This used to be like that in order to avoid a confusion
with the secondary station Ourense-San Francisco which is closed to
passenger traffic since 2020.

Fix the name, deactivate Ourense-San Francisco and make only the City as
suggestable to avoid a name confusion.

Also remove 6546:Orense as it probably came from an old SNCF dataset.