Move buses to Brugge main station
This commit introduces two changes:
- Add Bruges to the station of Brugge to make it more in line with what SNCB
  calls it on their website.
- Move bus codes from the meta station to the actual train station

The rationale being that the meta station is useful only when there is an
ambiguity in the stop names and to let customers choose the stops if needed.
There is not much benefit from it here for this relatively small city, esp.
considered that meta stations add complexity: the distinction between a
city and a stop probably flies over the head of many customers.