Refine Certosa LINK stations from Trenitalia
According to Trenitalia's API, the stations

* Polla (20704)
* Polla – Grotte di Pertosa-Auletta (74733)
* Certosa di Padula (74734)

share the same langitude & longitude values. However, this is not exactly like this in the real world.
According to the coach operator, a Certosa LINK ticket purchased with any of these location as the origin or destination of the journey will allow the passenger to travel up to any of them, no matter which one was chosen, for the same price.

After collecting the exact stop locations from the operator:

* Polla – Grotte di Pertosa-Auletta (74733) will point to the bus stop next to the Pertosa Caves: https://en.wikipedia.org/wiki/Pertosa_Caves
* Certosa di Padula (74734) will point to the bus parking nearby the Padula Charterhouse: https://en.wikipedia.org/wiki/Certosa_di_Padula