Fix Malpensa Aeroporto station mapping
This PR fixes the Malpensa Aeroporto station mapping for Distribusion.

The specific codes for Terminal 1 and Terminal 2 are not yielding results and are not used.
The Terminal 2 specific station code (used by all our connected coach carriers) has been fixed to the one that is actually returning results, while the Terminal 1 specific code has been removed because:

* it is the defaullt destination selected when using the area code, as it is closest to Milan
* the different enabled coach carriers enabled in Distribusion are using several, different station codes arount Terminal 1, so using the area is better and more maintenable.